sched_ule: tdq_setlowpri(): Remove useless argument
If not NULL, argument 'ctd' must exactly be 'tdq->tdq_curthread' and not
that, e.g., of a thread to add.  Moreover, the fact that it can be NULL
has to be treated with an additional test, which may be predicted wrong.

So, in addition to the possibility of passing the wrong thread (if,
e.g., some invariants in callers higher up in the chain are
inadvertently broken) because of the layer violation this argument
represents, we were essentially trading off reading a field in
a structure that is already or will have to be in the cache (and read
buffers) for a possible branch misprediction.  It's not worth it,
really.

Sponsored by:           The FreeBSD Foundation