POSIX.1b scheduling: libthr: Reject invalid parameters at setup
When some user tries to set per-thread scheduling parameters, at thread
creation via POSIX thread attributes or at run-time via
pthread_setschedparam() or pthread_setprio(), enforce that the passed
scheduling policy is known instead of silently treating it as
SCHED_OTHER.

Also, enforce that the value of 'sched_priority' is appropriate for the
scheduling policy (this is also checked on the kernel side, but doing so
here is necessary to avoid applying the priority transformation formula
to obtain a rtprio(2) priority), except for SCHED_OTHER since there
'sched_priority' is currently not used, and the corresponding rtprio(2)
priority used is always 0 (which anyway currently doesn't have any
influence, as will become apparent in subsequent commits).

Conversely, when requesting scheduling parameters, fail if the real
scheduling parameters are not representable with the supported POSIX.1b
scheduling policies.  No failure can normally happen when creating the
initial thread, where libthr asks the kernel about the current policy
via rtprio_thread(), because userland processes' internal scheduling
parameters are always translatable to POSIX.1b scheduling parameters, so
we just call PANIC() in this case.

While here, ensure that _thr_getscheduler() and _thr_setscheduler()
return an error value instead of -1 when then fail, simplifying the
logic in callers.  Rename 'ret' variables to 'error' there to reflect
this change.