POSIX.1b scheduling: thr_new() to accept POSIX priorities
This system call is used by 'libthr' to create a new thread and can be
passed a RT priority specification ('struct rtprio') that is applied
"atomically" at thread creation (that is, the thread immediately starts
with that priority applied, possibly causing it to preempt another
thread with lower priority and thus to run sooner than later).

Extend it with the possibility to pass POSIX.1b scheduling attributes
instead.

To preserve ABI (except in the case of weird programs that would set
previously unused bits in 'flags'), reserve bits in the 'flags' field to
version the system call.  The reserved bits are the whole second
most-significant byte (of the field), in order to reserve space for 256
versions (which should be more than enough but still can be extended)
and to avoid using the highest bit (an extra protection for old
misbehaving programs setting undefined flags, since 'flags' was signed
before this change).

Version 0 takes a 'struct rtprio' as before, while version 1 takes
scheduling attributes in the form of a 'struct sched_attr_v1'.  See also
prior commit "POSIX.1b scheduling: New system calls for thread
priorities" for more details on versioning of the related
thr_sched_set() and thr_sched_get() system calls.

More precisely, the 'rtp' field is replaced with 'sched_attr', also
a pointer (leaving the 'struct thr_param' layout unchanged) but to
unspecified data ('void *').  Code has to interpret the bits in 'flags'
corresponding to the THR_PF_VERSION_MASK mask to know if it points to
'struct rtp' or some version of 'struct sched_attr'.  It can do so via
the THR_PF_TO_VERSION() macro.  Another macro, THR_PF_FROM_VERSION(), is
provided for consumers to easily get the bits to or into 'flags' to
indicate a given version of 'struct sched_attr'.