RT priorities: Uniformize getting them
Rename pri_to_rtp() to rtp_get_thread().  Integrate into it the
p_cansee() check that all callers should be doing, simplifying most
callers' code.  With this change, rtp_get_thread() now returns an error
that callers must check.  Besides EPERM that can be returned by the
p_cansee() check, EOVERFLOW can be returned if translation from internal
scheduling priorities to a 'struct rtprio' is not possible (currently,
non-PRI_ITHD kernel threads minus those explicitly assigned a priority
in the range PRI_MIN_TIMESHARE to PRI_MAX_TIMESHARE currently).  We'll
augment the available RT priority types and fix the internal priority
used as a basis for the computation in subsequent commits so that
returning EOVERFLOW won't normally happen anymore.  This change in
passing fixes a bug where uninitialized kernel memory would be returned
in the case of a thread with interrupt priority class.

Create rtp_get_proc() that retrieves the RT priority for a whole
process, which is the maximum priority of all threads in the process.
This allows to remove duplicated code both in the implementation of
rtprio(2) and that dealing with ioprio in the Linux emulation layer.
The code comparing RT priorities was the same and wrong in both places.
It has now been fixed and only a single copy lives in the file-local
function _rtp_highest_pri().

sched_getscheduler(2) and sched_setparam(2) will now return EOVERFLOW on
RT priority types that can't be translated to POSIX(.1b) instead of
acting as if the corresponding class was SCHED_OTHER.  Again, later
additions should remove all impossible translation cases.