RT priorities: When setting, allow ID or pointer arguments
Split the functions doing the actual work for system calls rtprio() and
rtprio_thread(), i.e., kern_rtprio() and kern_rtprio_thread(), in
a variant taking an ID (as the system calls pass) and another which
already takes a pointer to a 'struct proc' or 'struct thread'.

The first variants are named with a '_by_id' suffix.  The second
variants reuse the original names, and require the process lock all the
way as a means to ensure that the thread structure isn't recycled while
the call is running, although technically it is strictly necessary only
to call rtp_{get,set}_{thread,proc}().  The rtprio_postamble() and
rtprio_thread_postamble() functions have been created to hold common
code for these variants.

The new functions will be used in subsequent commits (actually, most of
them but not all, which still are added here for completeness).

While here, include <sys/types.h> in <sys/syscallsubr.h> to have 'pid_t'
and 'lwpid_t' defined.  This wasn't strictly necessary since included
<sys/mount.h> already includes <sys/types.h>, but is much better
inclusion practice.