POSIX.1b scheduling: New system calls for thread priorities
These new system calls, thr_sched_set() and thr_sched_get(), allow to
set/retrieve priorities with the POSIX.1b semantics and format (and
might be extended in the future to allow other ones).

The reason for that change is the desire to hide from userland that the
implementation of POSIX.1b scheduling is built on top of the older, but
arguably more natural and currently mostly more feature rich, RT
priorities interface, with the following advantages:
- Remove the duplication in translation mechanisms (currently, they
exist both in kernel and in userland).
- Decouple kernel and userland, preventing the out-of-sync risk, e.g.,
in full jail scenarios.
- Enable (with other changes) to perform internal scheduling kernel
changes transparently to userland.

We also note that POSIX(.1b) has been defining system functions
sched_get_priority_min() and sched_get_priority_max() from the start,
which consumers must use to determine the allowable priority range per
scheduling class, allowing the kernel to prescribe a range that suits it
well with respect to other internal scheduling constraints.  Probably
for this reason, these two functions are already implemented as system
calls.  So, continue the logical trend that the kernel should be in
charge of which values are meaningful.  This doesn't exclude good
performance in case of repeated calls to these functions since it's easy
to add a userland cache for these values as a later step.

The new system calls operate on a structure holding all POSIX scheduling
attributes ('struct sched_attr'), i.e., currently, the policy and the
priority (the latter corresponding to 'sched_priority' of
'struct sched_param').  They have been designed carefully to allow
compatible evolution while minimizing the consumers' burden as well as
that of kernel maintainers, thanks to the following properties:
- These system calls take flags as their first parameter, including
  a version number.
- Users are provided with a separate structure definition for each
  version, to be used along with its corresponding version number.
  Version 1 is defined, and the corresponding C structure is
  'struct sched_attr_v1'.
- Future evolution will preserve the ABI, allowing to run older binaries
  on newer kernels, without the need to define new system calls.
- A kernel-internal version of 'struct sched_attr' is defined separately
  and conversion between this and userland version 1 happens in separate
  internal functions, whose implementation is "protected" by various
  compile-time assertions.

This design has been chosen over the simpler approach of defining only
system calls as needed now and leaving to the future the addition of new
versions of these principally to be consistent with the modifications we
are going to make to thr_new() in a subsequent commit.  Additionally, it
allows to evolve these system calls without cluttering 'syscalls.master'
with new definitions.  Although it requires that the object copied from
user memory appears untyped in the interface (parameter 'attr'), we
ensured that the signatures are compatible with SAL annotations by
introducing parameter 'len' which must be passed the number of bytes to
copy from 'attr'.

Note that these new system calls are not used yet by 'libthr', this will
happen in a subsequent commit.