libthr: Remove the hardcoded scheduling policies bounds
Instead, for the minimal and maximal priority per policy, directly use
sched_get_priority_min() and sched_get_priority_max(), which now rely on
a cache.

For the default priority to use when setting a policy via
pthread_attr_setschedpolicy(), for backwards compatibility, use the
minimum priority for SCHED_FIFO or SCHED_RR, or use the middle of the
available range for others.

In pthread_mutexattr_setprotocol(), change the PTHREAD_PRIO_PROTECT
mutexes' default ceiling to be the smallest priority, as was most
probably intended when the code was introduced (to be consistent with
pthread_attr_setschedpolicy()).

Next commit will get rid of this bogus default priority business, for
reasons explained there.  We do it in a separate commit because it is
debatable if that change can be MFCed, whereas this one should
definitely be.