POSIX.1b scheduling: Thread scheduling attributes: Sane defaults
TL;DR:
Don't reset parameters when setting the scheduling policy, because this
is simply not POSIX-compliant and violates POLA.  Opt for the simpler
and safer solution of initializing the scheduling policies and
parameters to invalid values, making it compulsory for users to call
pthread_attr_setschedpolicy() *and* pthread_attr_setschedparam() if they
set the 'inheritsched' thread attribute to PTHREAD_EXPLICIT_SCHED (else
thread creation will fail).  Similarly, prevent
pthread_mutexattr_setprotocol() from resetting the ceiling and instead
initialize it to an invalid value, which forces users to call
pthread_mutexattr_setprioceiling() when the protocol is set to
PTHREAD_PRIO_PROTECT.

With this change, _thr_sched_policy_default_priority() introduced a few
commits ago has no more uses, so remove it.  In
pthread_attr_setschedpolicy(), we instead validate the passed policy by
calling sched_get_priority_min() and returning the corresponding 'errno'
in case of error.

***

pthread_attr_setschedpolicy() is the only function in the POSIX API
where a scheduling policy can be specified without its parameters.  This
is because it operates on an opaque object holding thread attributes
(type 'pthread_attr_t').  The cause of this apparent peculiarity is that
the standard allows implementations to extend such objects with their
own attributes without changing the public interface.  To this end,
these objects' attributes are not mere fields in a C structure but are
rather hidden behind accessors, which the pthread_attr_*() functions
really are (except pthread_attr_init() and pthread_attr_destroy(),
obviously).

POSIX only says that a setter (accessor) changes the corresponding
attribute.  Moreover, since it doesn't prescribe any order between calls
to pthread_attr_setschedpolicy() and pthread_attr_setschedparam() (and,
generally, between all accessors), it follows that, necessarily,
pthread_attr_setschedpolicy() isn't allowed to change a previously-set
scheduling parameter.  Indeed, consider the following sequence of calls:
- pthread_attr_setschedparam() to set some policy's parameters.
- pthread_attr_setschedpolicy() to set the policy itself.
If pthread_attr_setschedpolicy() was allowed to change the parameters of
the policy, then the final scheduling parameters specified by the
attribute object would not be those intended by the program.

As expected, illumos, GNU/Linux, OpenBSD and NetBSD effectively don't do
that.  So let's stop doing that as well.

The standard leaves it up to the implementations to decide which default
values are set in the different attributes just after object creation.
Which policy should we adopt concerning all the scheduling attributes?

illumos, OpenBSD and NetBSD implement a simple answer: Priority is set
to 0 by default and is left unmodified (and the default policy is
SCHED_OTHER, if not explicitly set either).  It can be considered
deficient, since an explicitly priority may be invalid for the default
SCHED_OTHER policy (the converse is not true, since all these systems
admit 0 as a possible value for the POSIX policies, arguably in
violation of POSIX; more on this in a later commit that makes this
property false on FreeBSD).

GNU/Linux, i.e., glibc, gives a different answer: It uses internal flags
to track which attributes have been explicitly set, and, on
PTHREAD_EXPLICIT_SCHED at thread creation, fills the scheduling policy
and parameters that were not explicitly set with the creating thread's
current values (which realizes per-attribute inheritance).  Because, on
Linux, priority 0 is the only valid value for the SCHED_OTHER policy and
is not valid for every other policies, this choice appears worse than
the above-mentioned one.  Additionally, it borders on non-compliance
since the standard's model is that of default attribute values which are
then amended by users through the accessors, which may be considered to
rule out having an attribute's "default value" changing depending on how
others are set (in the most straightforward interpretation; but others
are possible that would admit this behavior, such as a default value
precisely indicating inheritance).

So we opt instead for a very simple, standard-compliant and safer
policy: We initialize the scheduling policy and priority to invalid
values.  That way, if PTHREAD_EXPLICIT_SCHED was requested, thread
creation will fail if the program doesn't call both
pthread_attr_setschedpolicy() *and* pthread_attr_setschedparam().  This
is the safest policy because it imposes really explicit attribute values
on PTHREAD_EXPLICIT_SCHED.  We strongly believe that this is what POSIX
should (have) standardize(d).

The same reasoning can be applied to the ceiling of PP mutexes (we have
not checked what the other implementations do for this one).