POSIX.1b scheduling: Define one invalid priority, check ranges
This constant is principally meant to be used internally in libthr to
indicate an unset/invalid attribute field.  It is in accordance with
POSIX, which rules out negative priority values.

Having an explicit constant in 'sys/sched.h' has the advantage of
allowing to modify the kernel to ensure that this value is never
accepted or returned as valid.  Given the current per-scheduling-class
priority ranges, it can't, but this is a cheap future-proof defense.

Note: The -1 value that SCHED_PRIORITY_INVALID expands to can be used as
an int, as it has to be in the POSIX interface, but also as a uint16_t
value in the internal storage ('struct sched_attr_v1').  This rules out
using 0xFFFF as a priority value in the implementation (in kernel and in
'libthr') as long as 'priority' stays of type uint16_t.