lib{c,sys}: Cache results of sched_get_priority_{min,max}()
The results of these system calls never change during a process'
lifetime (and, in fact, as long as a kernel with same ranges is used).

Use a relatively simple concurrent cache (no eviction) to avoid calling
them more than once per scheduling policy (unless there are concurrent
callers).

This cache is statically sized at 8 slots, which leaves some margin
above the number of currently (and soon-to-be) defined policies.  It
holds entries for each requested distinct policy, including invalid ones
(negative entries), until it becomes full, and maintains the invariant
that at most one slot can be assigned to a distinct policy value even in
the presence of concurrent callers.