POSIX.1b scheduling: Add LINUX_SCHED_NORMAL, use it
In the Linux kernel, the name of the usual policy, with value 0, is
SCHED_NORMAL, not SCHED_OTHER which is the name exported by glibc (and
mandated by POSIX as an interface to users).

Keep LINUX_SCHED_OTHER as an alias, but replace all occurences of it in
'linux_misc.c' by the proper name.