POSIX.1b scheduling: Process functions to operate on all threads
TL;DR:
Change POSIX's process scheduling attributes' accessor functions to
affect all threads in the target process, instead of the first process'
thread or the calling thread on PID 0.

In the Linux emulation, make sure to continue operating as before by
directly calling the kernel functions operating on threads (with the
proper thread).

***

POSIX says that threads with PTHREAD_SCOPE_SYSTEM scope (which is the
only scope we have actually supported since the KSE removal) are
scheduled according to their own attributes (those modified by
pthread_setschedparam() and pthread_setschedprio()) and not those of
their containing processes (those modified by sched_setscheduler() and
sched_setschedparam()), which should effectively be ignored.

At the same time, it also says that, at process creation, its first
thread's scheduling attributes are set to that of the process, and that
a just-created process (by fork() or posix_spawn() without the
POSIX_SPAWN_SETSCHEDPARAM or POSIX_SPAWN_SETSCHEDULER flags) inherits
the scheduling attributes of its parent when the latter's scheduling
policy is SCHED_FIFO or SCHED_RR.  All systems we have examined (see
below) work the same for SCHED_OTHER although POSIX leaves that behavior
implementation-defined.

Logically speaking, this combination of prescriptions seems to make
little sense: On the one hand, process scheduling attributes have no
effect for threads with PTHREAD_SCOPE_SYSTEM, but on the other hand the
first thread in a process inherits its attributes from it.

Unsurprisingly, none of the prominent open-source UNIX systems,
including GNU/Linux, illumos, OpenBSD, NetBSD and FreeBSD implement the
first requirement, i.e., that sched_setscheduler() and
sched_setschedparam() should have no effect, although these systems only
really support PTHREAD_SCOPE_SYSTEM (even if some accept
PTHREAD_SCOPE_PROCESS, with no effect).  On these calls, FreeBSD and
Linux change only the scheduling attributes of the process' main thread
or, if PID 0 is passed, those of the calling thread.  illumos and NetBSD
change the attributes of all the process' threads.  OpenBSD doesn't
implement these calls, and more generally the _POSIX_PRIORITY_SCHEDULING
option.

While Linux's behavior is understandable given the history of its thread
implementation (LinuxThreads was using clone() to create a process per
"thread"), the source of FreeBSD's is unclear.  Why the first thread and
not the others?  Why the calling thread on PID 0?  Code archeology seems
to answer "to allow per-thread operation with the same system call
normally dedicated to processes" to the second question and "because
it's easier while introducing thread support" to the first.

We argue that these process APIs are most likely used either by
single-threaded processes (in which case operating on a single thread or
the whole process is equivalent) or by administrators, who then
reasonably expect the whole process to be affected.  The only use we can
see for the single thread attributes modification approach is to support
single-threaded applications that would link to a library that is
threaded and uses high priority threads, preventing the application to
later interfere with threads created by the library.  If at all used,
this was a pretty niche use case back when pthread_setschedparam() and
pthread_setschedprio() started to be implemented in UNIX systems, and
more importantly we don't think it is relevant anymore since all
applications and libraries should now either completely control their
scheduling and not rely on libraries needing realtime or conversely not
messing with the priority of threads they don't own by just using the
pthread functions, more than 25 years after their standardization.
Public code searches (GitHub, Google) don't show any significant use of
sched_setscheduler() or sched_setparam().  In this light, the most
sensible policy is that these calls affect all threads in a process.

With this change, we join the ranks of illumos and NetBSD in
implementing what we see as the most logical, unsurprising approach.  It
is still in contradiction with POSIX, although closer to it in spirit
compared to the previous approach since now the whole process is really
affected.  Abiding by POSIX here would in practice render
sched_setscheduler() and sched_setparam() completely useless (since we
only support PTHREAD_SCOPE_SYSTEM, as said above), and we do not resign
ourselves to this waste of aptly-named system calls.

***

Additional notes

In linux_sched_setparam(), we could have chosen to call
kern_thr_sched_set() with SCHED_CURRENT, but since we perform
a translation of priority levels via posix_prio_from_linux() on
'linux_map_sched_prio' being set, and because this translation requires
knowing the current policy, we have to issue a call to
kern_thr_sched_get() first anyway.

In the process of coding this whole change, orphan the following
functions in preparation of centralizing all priority related functions
and removing 'ksched.c':
- getscheduler() (internal)
- ksched_setparam()
- ksched_getparam()
- ksched_setscheduler()
- ksched_getscheduler()