RT priorities: Require privilege to raise priority
The required privilege is PRIV_SCHED_RAISEPRIO, the same as already
required to decrease the per-process "nice" value.

When setting the priority of some process, the old priority of each of
its threads must be checked to see if it is lower than the new one, and
if so we must check that the calling thread has the appropriate
privilege.  Since having this privilege does not depend on the old or
new priority, it can be performed only once.  If assigning the new
priority doesn't lead to raising an existing thread's priority, no
privilege check is necessary.

Separate the conversion from the RT priority space to internal
priorities from actually setting the internal priority.  This makes it
possible to perform the conversion from the passed 'struct rtprio' once
as a preamble and apply the result to all threads (in the rtp_set_proc()
case), and also to more easily check whether applying the new priority
amounts to raising the current one.  This is also generally cleaner and
will ease later changes.

rtp_set_proc() now can return EAGAIN in addition to 0 or EPERM, if the
new RSP_ATOMIC_ON_NO_CONCURRENCY flag is passed in.  This change is
temporary.  Please see the comment before the function for more
explanations.  This currently impacts the return codes of all rtprio()
and rtprio_thread() calls with RTP_SET, and all calls to POSIX.1b's
process priority functions, i.e., sched_setscheduler() and
sched_setparam().  After the semantics for priority level in the
timeshare class are revised, the possibility to return EAGAIN will be
suppressed and the flag renamed.

The Linux emulation doesn't use the new flag, either explicitly in
linux_ioprio_set() or implicitly in the above-mentioned process priority
functions since it operates on the calling or main thread of the
process, and consequently calls rtp_set_thread() instead of
rtp_set_proc().  Thus, the behavior of its system calls is unchanged
(they can never return EAGAIN).