MAC/priority: Allow users in 'realtime' group to raise priority
As a follow-up of the previous commit, give to users of the realtime
group privilege to raise priority.  Without it, they would never be able
to leverage the privilege of setting realtime priorities by themselves.

Now, as expected, users in the 'idletime' group are not able to raise
the priority of their processes even if the latter are in the SCHED_IDLE
/ RTP_PRIO_IDLE priority class.