PRIV_SCHED_SETPRIORITY: Rename to PRIV_SCHED_RAISEPRIO
This privilege is going to be checked in subsequent commits when setting
rtprio(2) or POSIX.1b scheduling attributes, for two reasons:
1. This allows an administrator to confine threads/processes/users by
imposing a priority limit (that with which processes are launched).
2. This is a prerequisite to have priority levels in RTP_PRIO_NORMAL
correspond to nice values, since then privilege checks must be the same
as those in nice() and setpriority().

So rename the old privilege to reflect its extended usage.  Keep the old
name as an alias for compatibility with external code.