POSIX.1b scheduling: SCHED_OTHER now an alias of new SCHED_TIMESHARE
Same reasoning as in the previous commit ("RT priorities:
RTP_PRIO_NORMAL now an alias to new RTP_PRIO_TIMESHARE"), but with
differences.

SCHED_OTHER is mandated by POSIX, so is not just a compatibility alias,
and will never be removed.

For compatibility, it is important that SCHED_TIMESHARE and SCHED_OTHER
have the same value.  A program requesting some policy must get back the
exact same policy when asking for it, and for compatibility regular
processes should be seen as having the SCHED_OTHER policy.  Since there
is only a single scheduler-internal class to represent all these (and
these internal classes are bound to disappear), this imposes that
external values cannot be different.  The only drawback of this scheme
is if at some point we decided to change the meaning of SCHED_OTHER
(quite unlikely), which means changing its numerical value, then already
compiled old programs would still get SCHED_TIMESHARE when specifying
SCHED_OTHER (unless some compatibility layer is introduced).