PP mutexes: Check also for PRIV_SCHED_RAISEPRIO
At lock and unlock.  This essentially amounts to an additional privilege
check in the do_lock_pp() case, but the do_unlock_pp() is more complex.
There, if PRIV_SCHED_RTPRIO is present, PRIV_SCHED_RAISEPRIO must be
tested only if the new "inherited" priority (the one coming from another
PP mutex we hold) is higher (numerically lower) than the old one.
Priority inherited from PI mutexes is ignored since it is not subject to
realtime checks and can fluctuate.  Normally, since we are unlocking,
the new priority should never be higher than the old (inherited priority
can only lower on unlock, since it is the maximum of all PP mutexes'
ceilings).  But it is directly provided by userland, which could send
a higher value by mistake or maliciously, so has to be checked as well.