runq: Clarity and style pass
In runq_choose() and runq_choose_fuzz(), replace an unnecessary 'while'
with an 'if', and separate assignment and test of 'idx' into two lines.

Add missing parentheses to one 'sizeof' operator.

Remove superfluous brackets for one-line "then" and "else" branches (to
match style elsewhere in the file).

Declare loop indices in their 'for'.

Test for non-empty bit sets with an explicit '!= 0'.

Move TABs in some prototypes of <sys/runq.h> (should not split the
return type specifier, but instead separate the type specifier with the
function declarator).

No functional change intended.