runq: Hide function prototypes under _KERNEL
And some structure definitions as well.

This header really is not supposed to be included by userland, so should
just error in this case.  However, there is one remaining use for it in
a test: Getting the value of RQ_PPQ to ensure a big enough priority
level difference in order to guarantee that a realtime thread preempts
another.  This use will soon be obsoleted by guaranteeing that
a realtime thread always preempts another one with lower priority, even
if the priority level is very close.